Move person detection to new directory.
Move VWW project to new directory structure.
