+++ v0.1/benchmarks/visual_wake_words/trained_models/vww_96.h5
