+++ v0.1/benchmarks/visual_wake_words/trained_models/vww_96_int8.tflite
